feat: 로그인 시 약관에 동의하는 기능 추가
