+++ app/src/main/res/drawable/item_login_button.xml
@@ -1,0 +1,14 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <solid android:color="@color/white" />
+
+    <corners android:radius="15dp" />
+
+    <stroke
+        android:width="1.5dp"
+        android:color="@color/light_gray" />
+
+    <corners android:radius="15dp" />
+
+</shape>
