+++ app/src/main/res/layout/fragment_login.xml
@@ -32,12 +32,23 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/iv_logo" />
 
-    <com.google.android.gms.common.SignInButton
+
+    <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/btn_google_sign_in"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="16dp"
-        android:layout_marginTop="42dp"
+        android:layout_width="0dp"
+        android:layout_height="40dp"
+        android:layout_marginHorizontal="50dp"
+        android:layout_marginTop="50dp"
+        android:background="@drawable/item_login_button"
+        android:drawableStart="@drawable/ic_google"
+        android:fontFamily="@font/roboto_medium"
+        android:paddingHorizontal="8dp"
+        android:stateListAnimator="@null"
+        android:text="@string/sign_in_with_google"
+        android:textAlignment="center"
+        android:textAllCaps="false"
+        android:textColor="@color/black"
+        android:textSize="14sp"
         app:buttonSize="wide"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
